requirements.txt
python-dotenv
eth-ape
synthetix==0.1.10
silverback
